Update RenameComputer.ps1
Added some sensing if user doesn't want to change the name. Write to Key and Property to registry so next MDT step can query it for a decision on reboot or not. 
Also added part to sense if computer is already joined to domain. In this case Computer Rename is the only part needed.