Create AssignHostNameFromAD.ps1
This script will figure out the subnet it is on and based on it, it will search specific OU for a newly created computer Object (in the last 10 mins). Then assign that hostname. This is good when provisioning new computers for new users. This could be modified to use a list. 

*This works if the computer object was crated manually in the AD prior to deployment!